Markdown to docx RichText converter
This commits adds a function that takes in a dictionary of jinja2 tags
that are associated with Markdown objects. It converts a Markdown
object it to HTML using markdown, then it parses the HTML with
BeautifulSoup and adds RichText to a docx template using docxtpl's
DocxTemplate, based on the Jinja2 tags.

The function then returns the filled DocxTemplate.