Let 'lambda' appear in the names of variables
[Lambda expressions](https://docs.python.org/3/reference/expressions.html#lambda) need to have either
a parameter list (thus a space after the 'lambda') or be immediately followed by a colon. The regex used
by the `value()` function to search for lambdas would improperly throw an error if the name of the
variable being sought contained the word 'lambda' anywhere, even if it wasn't a python lambda.

This change corrects the regex in `value` to only throw an error when the word lambda is followed by
a space or a colon.