Add aria-labelledby to solo-fields on pages
If there is only one field on a page (like when using a separate `field`
attribute and something like `dropdown`), we know the best label for it is the
same label that the form gets, the #daMainQuestion element. While the
information might be redundant, in my opinion it's better than a select field
with no label at all.